Testbed coverage for winforms paths.
