--- changes/1964.removal.rst
@@ -1,1 +1,0 @@
-The location returned by ``toga.App.paths.app`` is now the folder that contains the file that defines the app class used by the app. If you are using a ``toga.App`` instance directly, this may alter the path that is returned.
